[REFACTOR] Move admin scss files into admin folder
