+++ src/client/react/admin/scss/_challenge-view.scss
@@ -1,4 +1,4 @@
-@import '../variables/bp';
+@import '../../scss/variables/bp';
 
 #dashboard-challenges {
 	h5 {
