+++ src/client/react/admin/scss/_dashboard.scss
@@ -1,4 +1,4 @@
-@import '../variables/bp';
+@import '../../scss/variables/bp';
 
 #admin-dashboard {
 	@media (min-width: $bp-s) {
