+++ src/client/react/admin/scss/_gamestate-view.scss
