+++ src/client/react/admin/scss/_markdown-preview.scss
@@ -1,4 +1,4 @@
-@import '../variables/fonts';
+@import '../../scss/variables/fonts';
 
 .markdown-preview {
 	.markdown.output.markdown-content, .markdown.output.help {
