+++ src/client/react/admin/scss/_navbar.scss
