+++ src/client/react/admin/scss/_team-profile.scss
