+++ src/client/react/admin/scss/_user-view.scss
@@ -1,4 +1,4 @@
-@import '../variables/bp';
+@import '../../scss/variables/bp';
 
 #dashboard-users {
 	.user-card {
