+++ src/client/react/admin/views/InstructionArticlesView/index.js
@@ -10,7 +10,7 @@ import InstructionArticleCard from './InstructionArticleCard';
 import InstructionArticleProfile from './InstructionArticleProfile';
 import FormInput from '../../../../../../lib/react/components/forms/FormInput';
 
-import '../../../scss/admin/_markdown-preview.scss';
+import '../../scss/_markdown-preview.scss';
 
 
 const QueryGetArticlesParams = '_id title created createdBy{username} modified modifiedBy{username}';
